test: reduce flakiness (#14383)
