+++ projects/storefrontapp-e2e-cypress/cypress/helpers/applied-promotions.ts
@@ -1,7 +1,7 @@
 import { waitForOrderToBePlacedRequest } from '../support/utils/order-placed';
 import { registerCartPageRoute } from './cart';
 import { verifyAndPlaceOrder } from './checkout-as-persistent-user';
-import { waitForPage, waitForProductPage } from './checkout-flow';
+import { waitForPage } from './checkout-flow';
 
 export const eosCameraProductName = 'EOS450D';
 
@@ -43,10 +43,8 @@ export function addProductToCart() {
   cy.wait(`@addToCart`);
 }
 
-export function goToCartDetailsViewFromCartDialog() {
-  cy.get('cx-added-to-cart-dialog').within(() => {
-    cy.findByText(/view cart/i).click();
-  });
+export function goToCartDetailsView() {
+  cy.get('cx-mini-cart').click();
 }
 
 export function selectShippingAddress() {
@@ -88,16 +86,7 @@ export function goToOrderHistoryDetailsFromSummary() {
 
 export function checkAppliedPromotions() {
   it('Should display promotions for product in cart and checkout', () => {
-    const eosCameraProductCode = '1382080';
-    const productPage = waitForProductPage(
-      eosCameraProductCode,
-      'getProductPage'
-    );
-    cy.visit(`/product/${eosCameraProductCode}`);
-    cy.wait(`@${productPage}`).its('response.statusCode').should('eq', 200);
-    addProductToCart();
-    checkForAppliedPromotionsInCartModal(eosCameraProductName);
-    goToCartDetailsViewFromCartDialog();
+    goToCartDetailsView();
     checkForAppliedPromotions();
 
     cy.get('.cart-details-wrapper > .cx-total').then(($cart) => {
@@ -138,6 +127,13 @@ export function removeCartEntry() {
   });
 }
 
+export function closeCartDialog() {
+  cy.get('cx-added-to-cart-dialog').should('be.visible');
+  cy.get('cx-added-to-cart-dialog').within(() => {
+    cy.get('button.close').click({ force: true });
+  });
+}
+
 export function checkAppliedPromotionsFordifferentCartTotals() {
   const batteryProductCode = '266685';
 
@@ -155,7 +151,8 @@ export function checkAppliedPromotionsFordifferentCartTotals() {
       )}/users/*/customercoupons`,
     }).as('customer_coupons');
 
-    goToCartDetailsViewFromCartDialog();
+    closeCartDialog();
+    goToCartDetailsView();
     cy.wait('@cart_page');
     cy.wait('@customer_coupons');
     cy.get('.cx-promotions').should('contain', '200');
